Update to use mythffmpeg instead of avconv
Debian (and Ubuntu) have switched back from the avconv fork to using
ffmpeg, meaning avconv is no longer available. Given mythtv bundles its
own version of ffmpeg, it makes sense to use that.

ffmpeg is being more strict when it comes to extracting the audio -
DVB-T recordings contain the audio in an mp2 track, not mp3. ffmpeg
errors when attempting to extract this to a .mp3 file, whereas avconv
didn't mind. mp3splt has no trouble reading a .mp2 file, so this updates
ffmpeg to use that extension.